Added support for JSOC's show_series wrapper
This also changes the Client.series() API. For more details, see below.

The show_series CGI at JSOC returns only a fraction of the available
data series. To get a full list of all available series, a wrapper
script called showextseries has to be called. Unfortunately, the wrapper
behaves differently than show_series and does not return any primekeys.

As a result the API of Client.series() was changed to return only a list
of series names by default. An additional option can be used to get the
full result, returning a DataFrame containing the columns name, primekeys
and info for show_series, or the columns name and info for the wrapper.