Changed the drms_json module into a package called drms
Previously, all code was contained in a single module file called
drms_json.py. This file was now split up into multiple submodules,
which were then combined to a package called drms.

The public interface was not altered by these changes, except that
the HttpJsonClient and HttpJsonExport classes are now "hidden" from
the global package namespace and can be only accessed by explicitly
importing the drms.json submodule.

In addition to the above changes, the setup script and all examples
were modified to account for the name change.