Improved examples running outside the source directory
Importing the example_helpers module now checks if the parent
directory contains the drms package and only adds it to the head
of the Python import path list, if the package was found. If the
drms package was not found in the parent directory, the examples
will instead try to import an already installed version.