Fixed a Python 2 import issue with versioneer
Versioneer 0.16 does not use absolute imports for Python 2. This
causes a problem with the short version of drms/_version.py, when
executed by Python 2, because this submodule tries to import the
Python json module, but gets the drms.json submodule instead.

The issue was fixed by adding absolute_import future imports to
the versioneer code and regenerating drms/_version.py.