Added email to Client constructor, changed export API
It is now possible to specify an export email address directly in
the Client constructor.

In addition the arguments of the export method were rearranged and
the email argument was moved further to the end of the argument
list. The export examples were modified to comply with the new
calling order of Client.export().