Client-side validation for some export protocol_args
There is currently no server-side validation of color tables and
color scaling for jpeg/movie export requests. Using invalid values
currently result in broken images/movies without any error message.
This is a workaround for the issue, until the server CGIs are fixed
and return proper errors in this case.

Note that the color table/scaling names must be updated in the
client code, whenever they change on the server. This data is
stored in the "jsoc.Color_Tables" series, which is only available
internally at Stanford.