Slightly changed license text to make GitHub happy
GitHub apparently does not identify the text in LICENSE.txt as MIT
license. This is probably caused by the additional line after the
Copyright notice that directs people to the AUTHORS.txt file. This
commit tries to fix this issue by removing the second line, and
adding a condensed version of its text to the first line.